WIP: moving to Qt 5, OS X first
* 'QChar' is deprecated, in a few places
* tests/auto/qcl:
FAIL!  : tst_QCL::argumentPassing() Compared floats are not the same (fuzzy compare)
   Actual   (buf[0]): 0
   Expected (3.0f)  : 3
   Loc: [/Users/liangqi/dev/git/qtopencl/tests/auto/qcl/tst_qcl.cpp(321)]
* tests/benchmarks/blur:
In file included from /Users/liangqi/dev/git/qtopencl/tests/benchmarks/blur/pixmapfilterwidget.cpp:42:
/Users/liangqi/dev/git/qtopencl/tests/benchmarks/blur/pixmapfilterwidget.h:46:10: fatal error: 'QtGui/private/qpixmapfilter_p.h' file not found
         ^
In file included from .moc/moc_pixmapfilterwidget.cpp:9:
.moc/../../../../../../dev/git/qtopencl/tests/benchmarks/blur/pixmapfilterwidget.h:46:10: fatal error: 'QtGui/private/qpixmapfilter_p.h' file not found
         ^
1 error generated.
make: *** [.obj/moc_pixmapfilterwidget.o] Error 1
make: *** Waiting for unfinished jobs....
1 error generated.
make: *** [.obj/pixmapfilterwidget.o] Error 1
In file included from /Users/liangqi/dev/git/qtopencl/tests/benchmarks/blur/tst_blur.cpp:56:
/Users/liangqi/dev/git/qtopencl/tests/benchmarks/blur/pixmapfilterwidget.h:46:10: fatal error: 'QtGui/private/qpixmapfilter_p.h' file not found
         ^
* demos/mandelbrot crash

+++ src/opencl/qcldevice.cpp
@@ -192,7 +192,7 @@ QCLDevice::DeviceTypes QCLDevice::deviceType() const
             != CL_SUCCESS)
         return QCLDevice::DeviceTypes(0);
     else
-        return QCLDevice::DeviceTypes(type);
+        return QCLDevice::DeviceTypes(static_cast<int>(type));
 }
 
 /*!
@@ -673,7 +673,7 @@ QCLDevice::FloatCapabilities QCLDevice::floatCapabilities() const
             != CL_SUCCESS)
         return NotSupported;
     else
-        return QCLDevice::FloatCapabilities(config);
+        return QCLDevice::FloatCapabilities(static_cast<int>(config));
 }
 
 /*!
@@ -693,7 +693,7 @@ QCLDevice::FloatCapabilities QCLDevice::doubleCapabilities() const
             != CL_SUCCESS)
         return NotSupported;
     else
-        return QCLDevice::FloatCapabilities(config);
+        return QCLDevice::FloatCapabilities(static_cast<int>(config));
 }
 
 /*!
@@ -713,7 +713,7 @@ QCLDevice::FloatCapabilities QCLDevice::halfFloatCapabilities() const
             != CL_SUCCESS)
         return NotSupported;
     else
-        return QCLDevice::FloatCapabilities(config);
+        return QCLDevice::FloatCapabilities(static_cast<int>(config));
 }
 
 /*!
